+++ app/src/main/java/pl/lambada/songsync/ui/screens/home/HomeScreen.kt
@@ -6,43 +6,21 @@ import androidx.compose.animation.AnimatedVisibilityScope
 import androidx.compose.animation.Crossfade
 import androidx.compose.animation.ExperimentalSharedTransitionApi
 import androidx.compose.animation.SharedTransitionScope
-import androidx.compose.animation.SizeTransform
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
-import androidx.compose.animation.slideInVertically
-import androidx.compose.animation.slideOutVertically
-import androidx.compose.animation.togetherWith
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.isImeVisible
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Clear
 import androidx.compose.material.icons.filled.Search
-import androidx.compose.material.icons.outlined.FilterAlt
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.ShapeDefaults
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
-import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -58,29 +36,23 @@ import androidx.compose.runtime.setValue
 import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.focus.onFocusChanged
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextRange
-import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import kotlinx.parcelize.Parcelize
-import pl.lambada.songsync.R
 import pl.lambada.songsync.data.MainViewModel
 import pl.lambada.songsync.domain.model.Song
 import pl.lambada.songsync.ui.ScreenAbout
 import pl.lambada.songsync.ui.ScreenSearch
 import pl.lambada.songsync.ui.screens.home.components.BatchDownloadLyrics
+import pl.lambada.songsync.ui.screens.home.components.FilterAndSongCount
 import pl.lambada.songsync.ui.screens.home.components.FiltersDialog
 import pl.lambada.songsync.ui.screens.home.components.HomeAppBar
+import pl.lambada.songsync.ui.screens.home.components.HomeSearchBar
+import pl.lambada.songsync.ui.screens.home.components.HomeSearchThing
 import pl.lambada.songsync.ui.screens.home.components.SongItem
 import pl.lambada.songsync.util.dataStore
 import pl.lambada.songsync.util.ext.BackPressHandler
@@ -185,9 +157,6 @@ fun LoadingScreen() {
     }
 }
 
-@Parcelize
-data class MyTextFieldValue(val text: String, val cursorStart: Int, val cursorEnd: Int) : Parcelable
-
 @OptIn(ExperimentalSharedTransitionApi::class)
 @Composable
 fun HomeScreenLoaded(
@@ -232,11 +201,10 @@ fun HomeScreenLoaded(
     Column {
         if (isBatchDownload) {
             BatchDownloadLyrics(
-                songs = if (selected.isEmpty()) displaySongs else songs.filter {
-                    selected.contains(
-                        it.filePath
-                    )
-                }.toList(),
+                songs = if (selected.isEmpty())
+                    displaySongs
+                else
+                    songs.filter { selected.contains(it.filePath) }.toList(),
                 viewModel = viewModel,
                 onDone = { onBatchDownload(false) })
         }
@@ -250,12 +218,7 @@ fun HomeScreenLoaded(
         ) {
             item {
                 Column(
-                    modifier = Modifier.padding(
-                        top = 5.dp,
-                        bottom = 5.dp,
-                        start = 22.dp,
-                        end = 4.dp
-                    ),
+                    modifier = Modifier.padding(top = 5.dp, bottom = 5.dp, start = 22.dp, end = 4.dp),
                 ) {
                     HomeSearchThing(
                         showingSearch = showingSearch,
@@ -330,139 +293,5 @@ fun HomeScreenLoaded(
     }
 }
 
-@Composable
-fun HomeSearchThing(
-    showingSearch: Boolean,
-    searchBar: @Composable () -> Unit,
-    filterBar: @Composable () -> Unit
-) {
-    AnimatedContent(
-        targetState = showingSearch,
-        transitionSpec = {
-            if (targetState) {
-                (slideInVertically { height -> height } + fadeIn()).togetherWith(
-                    slideOutVertically { height -> -height } + fadeOut()
-                )
-            } else {
-                (slideInVertically { height -> -height } + fadeIn()).togetherWith(
-                    slideOutVertically { height -> height } + fadeOut()
-                )
-            }.using(
-                SizeTransform()
-            )
-        },
-        label = "",
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(55.dp)
-    ) { showing ->
-        if (showing) searchBar() else filterBar()
-    }
-}
-
-@Composable
-fun HomeSearchBar(
-    query: TextFieldValue,
-    onQueryChange: (TextFieldValue) -> Unit,
-    showSearch: Boolean,
-    onShowSearchChange: (Boolean) -> Unit,
-    showingSearch: Boolean,
-    onShowingSearchChange: (Boolean) -> Unit,
-) {
-    val focusRequester = remember { FocusRequester() }
-    var willShowIme by remember { mutableStateOf(false) }
-
-    @OptIn(ExperimentalLayoutApi::class)
-    val showingIme = WindowInsets.isImeVisible
-
-    if (!showingSearch && showSearch) {
-        onShowingSearchChange(true)
-    }
-
-    if (!showSearch && !willShowIme && showingSearch && !showingIme && query.text.isEmpty()) {
-        onShowingSearchChange(false)
-    }
-
-    if (willShowIme && showingIme) {
-        willShowIme = false
-    }
-    val focusManager = LocalFocusManager.current
-
-    TextField(
-        value = query,
-        onValueChange = onQueryChange,
-        leadingIcon = {
-            Icon(
-                Icons.Filled.Search,
-                contentDescription = stringResource(id = R.string.search),
-                modifier = Modifier.clickable {
-                    onShowSearchChange(false)
-                    onShowingSearchChange(false)
-                }
-            )
-        },
-        trailingIcon = {
-            Icon(
-                imageVector = Icons.Default.Clear,
-                contentDescription = stringResource(id = R.string.clear),
-                modifier = Modifier.clickable {
-                    onQueryChange(TextFieldValue(""))
-                    onShowSearchChange(false)
-                    onShowingSearchChange(false)
-                }
-            )
-        },
-        placeholder = { Text(stringResource(id = R.string.search)) },
-        shape = ShapeDefaults.ExtraLarge,
-        colors = TextFieldDefaults.colors(
-            focusedLabelColor = MaterialTheme.colorScheme.onSurface,
-            focusedIndicatorColor = Color.Transparent,
-            unfocusedIndicatorColor = Color.Transparent,
-            disabledIndicatorColor = Color.Transparent
-        ),
-        modifier = Modifier
-            .padding(end = 18.dp)
-            .focusRequester(focusRequester)
-            .onFocusChanged {
-                if (it.isFocused && !showingIme) willShowIme = true
-            }
-            .onGloballyPositioned {
-                if (showSearch && !showingIme) {
-                    focusRequester.requestFocus()
-                    onShowSearchChange(false)
-                }
-            },
-        keyboardOptions = KeyboardOptions(imeAction = ImeAction.Search),
-        keyboardActions = KeyboardActions(
-            onSearch = { focusManager.clearFocus() }
-        )
-    )
-}
-
-@Composable
-fun FilterAndSongCount(
-    displaySongsCount: Int,
-    onFilterClick: () -> Unit,
-    onSearchClick: () -> Unit
-) {
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.Center
-    ) {
-        Text(text = "$displaySongsCount songs")
-        Spacer(modifier = Modifier.weight(1f))
-        IconButton(onClick = onFilterClick) {
-            Icon(
-                imageVector = Icons.Outlined.FilterAlt,
-                contentDescription = stringResource(R.string.search),
-            )
-        }
-
-        IconButton(onClick = onSearchClick) {
-            Icon(
-                imageVector = Icons.Default.Search,
-                contentDescription = stringResource(R.string.search),
-            )
-        }
-    }
-}
+@Parcelize
+data class MyTextFieldValue(val text: String, val cursorStart: Int, val cursorEnd: Int) : Parcelable
